Modified.

+++ App/InSituVis/InSituVis.h
@@ -8,5 +8,5 @@ extern "C" InSituVis::Adaptor* InSituVis_new();
 extern "C" void InSituVis_delete( InSituVis::Adaptor* self );
 extern "C" void InSituVis_initialize( InSituVis::Adaptor* self );
 extern "C" void InSituVis_finalize( InSituVis::Adaptor* self );
-extern "C" void InSituVis_put( InSituVis::Adaptor* self, float* values, int nvalues, int dimx, int dimy, int dimz );
-extern "C" void InSituVis_exec( InSituVis::Adaptor* self, float time_value, int time_index );
+extern "C" void InSituVis_put( InSituVis::Adaptor* self, double* values, int nvalues, int dimx, int dimy, int dimz );
+extern "C" void InSituVis_exec( InSituVis::Adaptor* self, double time_value, long time_index );
